+++ note_07-03-2025.md
@@ -247,7 +247,7 @@ export default {
 };
 </script>
 ```
-=> Component cha (ParentComponent) truyền vào nội dung (ở đây là một <h2> và một <p>), và nó sẽ được chèn vào vị trí của slot trong Card.
+=> Component cha (ParentComponent) truyền vào nội dung (ở đây là một ```html <h2> và một <p> ```), và nó sẽ được chèn vào vị trí của slot trong Card.
 ### Named Slots (Slot có tên)
 Named Slots cho phép xác định nhiều vị trí mà bạn có thể chèn nội dung vào trong component con. Khi sử dụng named slots, cần chỉ định tên cho mỗi slot để có thể dễ dàng truyền và nhận nội dung từ bên ngoài.
 
@@ -306,4 +306,55 @@ export default {
    + Trong ParentComponent.vue, chúng ta sử dụng cú pháp #header và #footer để chỉ định nội dung cho các slot có tên cụ thể.
    + Slot mặc định nhận nội dung giữa <Card></Card>.
 
+### Scoped Slots (Slot có phạm vi)
+Scoped Slots là một loại slot đặc biệt cho phép component con truyền dữ liệu ngược lại cho component cha. Dữ liệu này có thể được sử dụng trong nội dung của slot trong component cha. Scoped slots rất hữu ích khi muốn truyền các thông tin từ component con mà không làm thay đổi cấu trúc của component cha.
 
+- Component con (ChildComponent.vue):
+```vue
+<template>
+  <div>
+    <slot :message="message"></slot>  <!-- Scoped slot, truyền giá trị 'message' cho component cha -->
+  </div>
+</template>
+
+<script>
+export default {
+  name: 'ChildComponent',
+  data() {
+    return {
+      message: 'Hello from Child!'
+    };
+  }
+};
+</script>
+
+
+```
+- Component cha (ParentComponent.vue):
+```vue
+<template>
+  <div>
+    <ChildComponent>
+      <!-- Sử dụng scoped slot, và nhận dữ liệu từ 'message' -->
+      <template #default="{ message }">
+        <p>{{ message }}</p>  <!-- Hiển thị dữ liệu 'message' từ ChildComponent -->
+      </template>
+    </ChildComponent>
+  </div>
+</template>
+
+<script>
+import ChildComponent from './ChildComponent.vue';
+
+export default {
+  components: {
+    ChildComponent
+  }
+};
+</script>
+
+
+```
+
+=> +  Trong ChildComponent, slot được khai báo với một thuộc tính :message="message". Điều này có nghĩa là component con sẽ truyền dữ liệu message ra ngoài cho component cha.
+   +  Trong ParentComponent, chúng ta sử dụng cú pháp #default="{ message }" để nhận giá trị của message từ slot trong ChildComponent và hiển thị nó trong template của component cha.
