+++ note_10-03-25.md
@@ -1,0 +1,186 @@
+# Reactivity & Computed Properties
+
+## Reactive Data (reactive, ref)
+
+- reactive là một hàm được dùng để tạo ra đối tượng phản ứng (reactive object) trong Vue 3, nghĩa là khi giá trị của đối tượng thay đổi, giao diện người dùng sẽ tự động cập nhật.
+
+```js
+import { reactive } from "vue";
+
+const state = reactive({
+  count: 0,
+  message: "Hello Vue 3!",
+});
+
+// Khi thay đổi giá trị, Vue tự động cập nhật giao diện
+state.count++;
+state.message = "Updated Message";
+```
+
+```js
+<template>
+  <div>
+    <p>Count: {{ state.count }}</p>
+    <button @click="increment">Increment</button>
+  </div>
+</template>
+
+<script>
+import { reactive } from 'vue'
+
+export default {
+  setup() {
+    const state = reactive({
+      count: 0
+    })
+
+    function increment() {
+      state.count++
+    }
+
+    return { state, increment }
+  }
+}
+</script>
+```
+
+=> Trong ví dụ trên, state.count là một reactive data. Khi giá trị của count thay đổi, Vue sẽ tự động cập nhật giao diện.
+
+- ref dùng để tạo ra các giá trị đơn lẻ phản ứng (reactive). ref đặc biệt hữu ích khi muốn quản lý dữ liệu kiểu cơ bản như số, chuỗi, hoặc boolean.
+- Khi sử dụng ref, cần truy cập giá trị thông qua thuộc tính .value
+
+```js
+import { ref } from "vue";
+
+const count = ref(0);
+
+// Đọc giá trị
+console.log(count.value); // 0
+
+// Thay đổi giá trị
+count.value++;
+console.log(count.value); // 1
+```
+
+```js
+<template>
+  <div>
+    <p>Count: {{ count }}</p>
+    <button @click="increment">Increment</button>
+  </div>
+</template>
+
+<script>
+import { ref } from 'vue'
+
+export default {
+  setup() {
+    const count = ref(0)
+
+    function increment() {
+      count.value++
+    }
+
+    return { count, increment }
+  }
+}
+</script>
+```
+
+- Trong template, Vue tự động "unwrap" các ref nên bạn không cần sử dụng .value:
+
+```html
+<template>
+  <div>Số lượng: {{ count }}</div>
+  <button @click="count++">Tăng</button>
+</template>
+```
+
+## Computed Properties
+
+- Viết biểu thức trực tiếp trong template rất tiện, nhưng chỉ dành cho những biểu thức có tính toán đơn giản. Những biểu thức phức tạp được viết theo cách đó sẽ khiến template cồng kềnh và khó bảo trì. Ví dụ:
+
+```html
+<div id="example">{{ message.split('').reverse().join('') }}</div>
+```
+
+=> Đó là lí do tại sao đối với bất kì logic nào phức tạp, bạn nên sử dụng
+computed property. - omputed là một tính năng cho phép tạo ra các giá trị tính
+toán dựa trên reactive data. Các giá trị tính toán này chỉ được tính lại khi dữ
+liệu phụ thuộc thay đổi. Điều này giúp tối ưu hiệu suất và tránh việc tính toán
+lại mỗi lần render.
+
+```vue
+<template>
+  <div>
+    <p>{{ reversedMessage }}</p>
+  </div>
+</template>
+
+<script>
+export default {
+  name: 'ComputedProperty',
+  data() {
+    return {
+   message: 'người đông bến đợi thuyền xuôi ngược'
+    };
+  }
+  computed: {
+     // một computed getter
+    reversedMessage() {
+      return this.message.split('').reverse().join('');
+    }
+  }
+};
+</script>
+```
+
+### Computed caching và phương thức:
+
+- Thay vì sử dụng computed property, chúng ta cũng có thể dùng một phương thức thay thế. Nếu xét về kết quả cuối cùng
+  thì hai cách tiếp cận này thât ra chỉ là một. Tuy nhiên, sự khác biệt ở đây là
+  computed property được cache lại dựa vào những những thành phần phụ thuộc
+  (dependency). Một computed property chỉ được tính toán lại khi những thành phần
+  phụ thuộc của chúng thay đổi. Điều này có nghĩa: miễn là giá trị của message
+  không thay đổi, thì những truy cập tới computed reversedMessage sẽ ngay lập tức
+  trả về kết quả được tính toán trước đó mà không phải chạy lại hàm một lần nữa.
+
+### Computed Setter
+
+- Những computed property mặc định chỉ có getter, nhưng bạn cũng có thể cung cấp setter nếu cần thiết:
+
+```vue
+<template>
+  <div>
+    <p>{{ fullName }}</p>
+    <p>{{ fullName("Viet Hoang") }}</p>
+  </div>
+</template>
+
+<script>
+export default {
+  name: 'ComputedProperty',
+  data() {
+    return {
+   firstName: 'John',
+   lastName: 'Doe'
+    };
+  }
+  computed: {
+     // một computed getter
+    fullName() {
+      return this.firstName + ' ' + this.lastName
+    }
+    // một computed setter
+    fullName(newValue) {
+      const names = newValue.split(' ')
+        this.firstName = names[0]
+        this.lastName = names[names.length - 1]
+    }
+  }
+};
+//Sử dụng setter
+</script>
+```
+
+## Watchers
