+++ src/components/Header/Header.vue
@@ -1,0 +1,16 @@
+<template>
+  <header class="header">Header</header>
+</template>
+
+<script>
+export default {
+  name: "HeaderComponent",
+};
+</script>
+<style scoped>
+.header {
+  font-size: 2.5rem;
+  padding: 20px;
+  text-align: center;
+}
+</style>
