+++ src/main.js
@@ -1,11 +1,10 @@
 import Vue from "vue";
 import App from "./App.vue";
+import router from "./router/router";
 
 Vue.config.productionTip = false;
 
 new Vue({
-  data: {
-    message: "hello",
-  },
   render: (h) => h(App),
+  router,
 }).$mount("#app");
