+++ src/views/About/About.vue
@@ -1,0 +1,9 @@
+<template>
+  <div>About</div>
+</template>
+
+<script>
+export default {
+  name: "AboutPage",
+};
+</script>
