+++ src/views/Home/Home.vue
@@ -1,0 +1,9 @@
+<template>
+  <div>Home</div>
+</template>
+
+<script>
+export default {
+  name: "HomePage",
+};
+</script>
